move a bunch of styles around
