+++ frontend/src/StateRepDetails.tsx
@@ -5,7 +5,7 @@ import {
   StateRepresentativeSponsorship
 } from './types';
 import React, { useState } from 'react';
-import styles from './style/PersonDetailsPanel.module.scss';
+import styles from './style/components/PersonDetailsPanel.module.scss';
 import useMountEffect from '@restart/hooks/esm/useMountEffect';
 import useApiFetch from './useApiFetch';
 import { Link } from 'react-router-dom';
