+++ frontend/src/style/App.scss
@@ -1,4 +1,4 @@
-// Global styles for the whole app. Use sparingly.For styles specific to the
+// Global styles for the whole app. Use sparingly. For styles specific to the
 // root page, use App.module.scss instead.
 //
 // The main styling stack is SCSS + CSS modules for each component, with
