+++ frontend/src/style/components/BillListItem.module.scss
@@ -1,4 +1,4 @@
-@import './_variables.scss';
+@import '../_variables.scss';
 
 .itemContainer {
     display: grid;
