+++ frontend/src/style/components/NotificationSettingsPanel.module.scss
@@ -1,4 +1,4 @@
-@import './_variables.scss';
+@import '../_variables.scss';
 
 .title {
     padding-top: 1.4rem;
