+++ frontend/src/style/components/PersonDetailsPanel.module.scss
@@ -1,4 +1,4 @@
-@import './_variables.scss';
+@import '../_variables.scss';
 
 .root {
     display: grid;
