+++ frontend/src/style/components/StateBillDetails.module.scss
