+++ frontend/src/style/components/UserList.module.scss
@@ -1,4 +1,4 @@
-@import './_variables.scss';
+@import '../_variables.scss';
 
 .container {
   display: grid;
