+++ frontend/src/style/pages/BillDetailsPage.module.scss
@@ -1,4 +1,4 @@
-@import './_variables.scss';
+@import '../_variables.scss';
 
 .title {
     color: $palette-medium-blue;
