+++ frontend/src/style/pages/RequestLoginLinkPage.module.scss
@@ -1,4 +1,4 @@
-@import './_variables.scss';
+@import '../_variables.scss';
 
 .fullScreenContainer {
   width: 100%;
