+++ frontend/src/style/pages/SettingsPage.module.scss
@@ -1,4 +1,4 @@
-@import './_variables.scss';
+@import '../_variables.scss';
 
 .title {
   color: $palette-medium-blue;
@@ -11,12 +11,14 @@
 
 .content {
   padding: 20px;
-}
 
-.searchBox {
-  margin-bottom: 1rem;
+  > h2 {
+    font-size: 1.4rem;
+    font-family: $title-font;
+    color: $palette-medium-blue;
+  }
 
-  ::placeholder {
-    font-style: italic;
+  > p {
+    font-size: 0.9rem;
   }
 }
